requirements.txt
streamlit>=1.30.0
sxtwl>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
